controller and services

+++ documentmanagement_documentmanagement/src/main/java/com/smartai/documentmanagement/serviceImpl/UserServiceImpl.java
@@ -3,8 +3,10 @@ package com.smartai.documentmanagement.serviceImpl;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import com.smartai.documentmanagement.ResourceNotFoundException;
 import com.smartai.documentmanagement.entity.User;
 import com.smartai.documentmanagement.repository.UserRepository;
 import com.smartai.documentmanagement.service.UserService;
@@ -15,8 +17,10 @@ import lombok.AllArgsConstructor;
 public class UserServiceImpl implements UserService{
 
     public final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
     @Override
     public User save(User user) {
+       user.setPassword(passwordEncoder.encode(user.getPassword()));
        return userRepository.save(user);
     }
 
@@ -44,6 +48,9 @@ public class UserServiceImpl implements UserService{
 
     @Override
     public void deleteUser(Long id) {
+    	if (!userRepository.existsById(id)) {
+            throw new ResourceNotFoundException("User with ID " + id + " not found");
+        }
         userRepository.deleteById(id);
     }
